Base done, lots of mining.
My base is complete.  It's a two story plus attic, plus basement. The garden,
dug out of the westerly end of the island, grows carrots and wheat. Cocoa beans
from an unexpected trip to the north grow on the lawn.  I've also built an
offshore tree farm.

Beneath the basement, I've dug a pit mine, straight down almost to bedrock.
From there, I dug diamond shafts, and found SO MANY DIAMONDS.

Tomorrow, I leave for the land to the east.