After a long spelunk.
While on my way to the new land, I noticed a spot along the shore where
the water was falling into the ground.  I swam down, and found a deep,
long cave that I eventually found to interesct two underground ravines.

After spelunking for an afternoon, and loading up with as much loot and
ores as I could find, I returned to the surface, back through the
opening in the shore.  I returned home with my findings, richer 64
blocks of coal, 20+ blocks of iron and redstone, a couple of emeralds,
and two diamonds.

I also decided on a place for a cliffside fortress.  It will be built
right into the cliff, protruding at the base, and conforming to the old
surface above.  it shall be glorious, and a perfect place to beging the
quest for enough riches to attempt a foray to the Nether.

sd.